add esLint config, rename components, add projects and registry, update general structure
