+++ src/components/footer/Footer.jsx
@@ -3,21 +3,21 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles((theme) => ({
   footer: {
     backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(9),
-    height: 'auto'
+    height: '200px',
+    position: 'fixed',
+    width: '100%',
   },
 }));
 
 export default function Footer() {
-
   const classes = useStyles();
 
   return (
 
-    <React.Fragment>
+    <>
       <CssBaseline />
       <footer className={classes.footer}>
         <Typography variant="h6" align="center" gutterBottom>
@@ -27,6 +27,6 @@ export default function Footer() {
           Something here to give the footer a purpose!
         </Typography>
       </footer>
-    </React.Fragment>
+    </>
   );
 }
