--- src/components/pages/main/Main.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import Header from './mainHeader/MainHeader'
-import MainContainer from './mainContainer/MainContainer'
-import Footer from './mainFooter/MainFooter'
-
-
-
-export default function Main() {
-
-  return (
-    <div>
-      <Header />
-      <MainContainer />
-      <Footer />
-    </div>
-  );
-}
